Add book implementation: Genre field has been added, books can be both added and removed, books can be searched and filtered from the book list. (Fix 2: modified according to the feedbacks)
