+++ src/main/java/com/bookclub/model/Genre.java
@@ -23,8 +23,4 @@ public enum Genre {
     public String toString() {
         return displayName;
     }
-
-
-
-
 }
